requirements.txt
h5py==2.9.0
matplotlib==3.1.0
numpy==1.16.4
pandas==0.25.0
Pillow==6.1.0
PyYAML==5.1.1
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.0
tensorboard==1.14.0
tensorflow-gpu==1.14.0
tensorflow-estimator==1.14.0
